Backend: Fix concurrency model
Add a mutex to the session to prevent race conditions
