+++ backend/api_media_attachments.go
@@ -31,7 +31,7 @@ func api_addMediaAttachment(response http.ResponseWriter, request *http.Request)
 		return
 	}
 
-	if !session.write {
+	if !session.CanWrite() {
 		ReturnAPIError(response, 403, "ACCESS_DENIED", "Your current session does not have permission to make use of this API.")
 		return
 	}
@@ -262,7 +262,7 @@ func api_updateMediaAttachment(response http.ResponseWriter, request *http.Reque
 		return
 	}
 
-	if !session.write {
+	if !session.CanWrite() {
 		ReturnAPIError(response, 403, "ACCESS_DENIED", "Your current session does not have permission to make use of this API.")
 		return
 	}
@@ -369,7 +369,7 @@ func api_removeMediaAttachment(response http.ResponseWriter, request *http.Reque
 		return
 	}
 
-	if !session.write {
+	if !session.CanWrite() {
 		ReturnAPIError(response, 403, "ACCESS_DENIED", "Your current session does not have permission to make use of this API.")
 		return
 	}
