+++ backend/api_media_audios.go
@@ -29,7 +29,7 @@ func api_addMediaAudioTrack(response http.ResponseWriter, request *http.Request)
 		return
 	}
 
-	if !session.write {
+	if !session.CanWrite() {
 		ReturnAPIError(response, 403, "ACCESS_DENIED", "Your current session does not have permission to make use of this API.")
 		return
 	}
@@ -306,7 +306,7 @@ func api_removeMediaAudioTrack(response http.ResponseWriter, request *http.Reque
 		return
 	}
 
-	if !session.write {
+	if !session.CanWrite() {
 		ReturnAPIError(response, 403, "ACCESS_DENIED", "Your current session does not have permission to make use of this API.")
 		return
 	}
@@ -397,7 +397,7 @@ func api_renameMediaAudioTrack(response http.ResponseWriter, request *http.Reque
 		return
 	}
 
-	if !session.write {
+	if !session.CanWrite() {
 		ReturnAPIError(response, 403, "ACCESS_DENIED", "Your current session does not have permission to make use of this API.")
 		return
 	}
