+++ backend/api_media_extended_description.go
@@ -28,7 +28,7 @@ func api_setExtendedDescription(response http.ResponseWriter, request *http.Requ
 		return
 	}
 
-	if !session.write {
+	if !session.CanWrite() {
 		ReturnAPIError(response, 403, "ACCESS_DENIED", "Your current session does not have permission to make use of this API.")
 		return
 	}
