+++ backend/api_media_thumbnail.go
@@ -27,7 +27,7 @@ func api_editMediaThumbnail(response http.ResponseWriter, request *http.Request)
 		return
 	}
 
-	if !session.write {
+	if !session.CanWrite() {
 		ReturnAPIError(response, 403, "ACCESS_DENIED", "Your current session does not have permission to make use of this API.")
 		return
 	}
