+++ backend/api_tags.go
@@ -67,7 +67,7 @@ func api_tagMedia(response http.ResponseWriter, request *http.Request) {
 		return
 	}
 
-	if !session.write {
+	if !session.CanWrite() {
 		ReturnAPIError(response, 403, "ACCESS_DENIED", "Your current session does not have permission to make use of this API.")
 		return
 	}
@@ -179,7 +179,7 @@ func api_untagMedia(response http.ResponseWriter, request *http.Request) {
 		return
 	}
 
-	if !session.write {
+	if !session.CanWrite() {
 		ReturnAPIError(response, 403, "ACCESS_DENIED", "Your current session does not have permission to make use of this API.")
 		return
 	}
